Start building styles, and modifying overrides
For this I had to remove the styles that were added in 139fd2c to alerts. I'm waiting about the rationalization for these, and then I'll see if I can make a workaround. At the moment it meant that alerts weren't showing on white backgrounds.
Added a page that display the styles without any tweaks necessary. This is for issue #85.