Update application.html.erb
meta-tag addition seo-stuff
open-graph support also g+ covering

!Some additions necessary to uncomment some tags, also dynamical content as i would suggest!

Added some missing meta tags to work with, some are filled by me, some are commented out to get filled by qualified person.
For some tags dynamical created content should be provided as describet by placeholders

FB Meta Tags for connecting the web-page to a facebook page via fb-app, this is the normal way to do it.
APP-ID and APP-ADMIN-ID nedded to provide this.

Open-Graph support is adding some usefull tags to have a better look on facebook while people are linking to the website.
Open Graph would use the base description tag if not provided and just takes any pictur available.
With the Open-Graph Tags implemented, one could use it to manage diffrent content to fb then to serach-engines
Most important feature is giving explicit images where the fb-user can choose from while linking. This could be helpfull for new features, announcements and so on.
IMPORTANT!
To refresh the fb-open-graph directly and to make sure everything works well, the following link will force a refresh of the open-graph data when the url is requested, the page then will show what is in their cache.