fix title on edit guide page to say 'Edit' instead of 'Create'
The same template (_new_guide_steps.html.erb) was being used for creating a
new guide AND for editing an existing guide. To fix this, a local variable
'edit' is passed to the template when it is rendered for editing. When
edit = true, a different title is given which says 'Edit A Growing Guide'