Move directly to crop page when user clicks on typeahead search.
Uses a somewhat roundabout method to avoid duplicating route information
into the static javascript pages - the view generates a path template that
it passes to the static JS.

The js-routes gem seems it might provide a cleaner approach, but it came
with a series of caveats about production problems. Pragmatism and all that,
this seemed like a useful way to do it - ceveat that I've never used rails,
so I may be missing something obvious.