Increase logo size and make it clickable everywhere
Make the logo clickable from everywhere it is displayed. Now, the logo refreshes the home page when clicked from the home page (and still redirects to the home page when clicked from elsewhere). Increase the logo size to 210x41 and add margin: 20px 10px 10px 30px. Increase the top bar height to accomodate the taller logo. Vertically align the other elements in the re-sized top bar so that the bottom of each element lines up with the bottom of the text in the logo.