Add margin-left: unset to logo-and-terms links
Links in the footer have a margin of -0.5rem which ensures each link
lines up under it's appropriate heading nicely. However, the "Terms of
Service" and "Privacy" links are different in that they are not aligned
in colums and are separated by an ampersand.

The spacing between the ampersand the the word "Privacy" is not the
same as between the end of "Terms of Service" and the start of the ampersand.
This is because "Privacy" is pulled left by the -0.5rem. Removing the margin
for these two links evens out the spacing.